Make timeout description more clear
Since timeout didn't work with animations as mentioned in #255, combined with the unclear documentation and only a boolean being shown, I was under the impression that this option only accepted booleans. I didn't even know we could specify a timeout time until I saw the sample code in that issue.