Enable undocumented 'within'; pass options to layouts
Enable the undocumented `within` option

This commit does two major things:
1) It passes all options through to the `style` method in layouts
2) It allows notys on custom containers to use a layout besides `inline`

Change 2 facilitates change 1: By allowing any layout on custom
containers, it allows the layout to decide a few details about which
way to display.

Specifically, if the `within` option is truthy, the position of the
noty is `absolute` instead of `fixed`. This allows notys inside of a
container to be calculated based on that container instead of the
`window`. One important caveat is that the container having the noty
assigned to it must be relative to trigger CSS' default
relative > absolute positioning.

Related, the `center` positioned elements also use `within` to decide
whether to use `$(this).parent()[0]` or `window` to do centering
calculations.