Fix absolute positioning bug
Fix absolute positioning bug

If a custom element is used, but `within` is not truthy, the behavior
should remain the same (use the `inline` layout).

This commit fixes a regression that was causing notys within custom
containers that were not instantiated with `within` as truthy to be
absolutely positioned at the top of the window.