Update confirm.md
Setting our variable n to an instance of new Noty({ ...}).show() throws a Typescript error on the invocation of n.close() as .close should not exist on type void (from the Class declaration line 11: show: () => void;). Instead set n to a new Noty() and use the .show() method outside of the variable n declaration.